requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0

# OpenAI
openai==1.51.0

# DB / SQL Server
pyodbc==5.2.0
sqlalchemy==2.0.35

# ML
scikit-learn
joblib
numpy

# Local STT (Whisper)
openai-whisper

# Env
python-dotenv==1.0.1

# AWS S3
boto3>=1.35.0

